+++ inst/GerminaQuant/pkgs.R
@@ -1,0 +1,22 @@
+# -------------------------------------------------------------------------
+# GerminaR-----------------------------------------------------------------
+# -------------------------------------------------------------------------
+#> open https://flavjack.github.io/GerminaR/
+#> open https://flavjack.shinyapps.io/germinaquant/
+#> author .: Flavio Lozano-Isla (lozanoisla.com)
+#> date .: 2020-04-20
+# -------------------------------------------------------------------------
+
+suppressPackageStartupMessages({
+  library(GerminaR)
+  library(shiny)
+  library(metathis)
+  library(tidyverse)
+  library(shinydashboard)
+  library(shinyWidgets)
+  library(gsheet)
+  library(readxl)
+  library(cowplot)
+})
+
+